Fix constant.

+++ lasagna-master/gallen/lasagna_master.go
@@ -25,7 +25,7 @@ func AddSecretIngredient(friendsList, myList []string) []string {
 func ScaleRecipe(amounts []float64, portions int) []float64 {
     var result []float64
     // Simplification: v/2 * portions  ===  v * 1/2 * portions === v * portions/2
-    var factor = float64(portions)/float64(2)
+    var factor = float64(portions)/2.0
     for _, v := range amounts {
         result = append(result, v * factor)
     }
